registerAPI/requirements.txt
flask
flask_restful
requests
mysql-connector-python